Fix #424: ckp
